+++ scooter/scooter/cites-project/model/src/main/java/isel/sisinf/model/interfaces/ITopUp.java
@@ -2,6 +2,7 @@ package isel.sisinf.model.interfaces;
 
 import isel.sisinf.model.Card;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public interface ITopUp {
@@ -12,6 +13,6 @@ public interface ITopUp {
     public Card getCard();
     public void setCard(Card card);
 
-    public double getValue();
-    public void setValue(double value);
+    public BigDecimal getValue();
+    public void setValue(BigDecimal value);
 }
